sft/requirements.txt
torch>=2.4.0
transformers>=4.51.0
accelerate>=0.34.0
datasets>=2.20.0
peft>=0.12.0
safetensors>=0.4.3
pyyaml>=6.0
